requirements.txt
# FastAPI framework for building APIs
fastapi==0.100.0  # Replace with the latest compatible version0  # Replace with the latest compatible version

# ASGI server for running FastAPI
uvicorn==0.22.0

# SQLAlchemy ORM for database interactions
sqlalchemy==1.4.46

# MySQL driver for SQLAlchemy
pymysql==1.0.3

# Environment variable management
python-dotenv==1.0.0

# PDF text extraction
pdfminer.six==20221105  # Compatible with pdfplumber
pdfplumber==0.10.0  # For extracting text from PDFs

# DOCX text extraction
python-docx==0.8.11

# Rate limiting for API calls
ratelimit==2.2.1

# Data validation and settings management
pydantic==2.3.0

# PyMuPDF for extracting images from PDFs
pymupdf==1.24.11

# Handling file uploads in FastAPI
python-multipart==0.0.5

# JWT authentication
pyjwt==2.8.0

# Email validation
email-validator==2.2.0

# Image processing
Pillow==9.5.0  # Downgraded to satisfy streamlit's requirement

# Streamlit for building the frontend
streamlit==1.24.0

# HTTP requests
requests==2.31.0

# Regular expressions for text processing
regex==2023.6.3